Rename user-dto file
