+++ src/auth/dto/forgot-password-response.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from "@nestjs/swagger";
-import { UserListDto } from "src/users/dto/user-list.dto";
+import { UserListDto } from "src/users/dto/user.dto";
 
 export class ForgotPasswordResponseDto {
   @ApiProperty()
