+++ src/auth/dto/login-response.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { UserListDto } from 'src/users/dto/user-list.dto';
+import { UserListDto } from 'src/users/dto/user.dto';
 import { UserDto } from './user.dto';
 
 export class LoginResponseDto {
