+++ src/auth/dto/reset-password-response.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { UserListDto } from '../../users/dto/user-list.dto';
+import { UserListDto } from '../../users/dto/user.dto';
 
 export class ResetPasswordResponseDto {
   @ApiProperty()
