+++ src/reports/entities/report.submission.entity.ts
@@ -4,6 +4,7 @@ import {
   Column,
   ManyToOne,
   OneToMany,
+  Index,
   JoinColumn,
 } from "typeorm";
 import { User } from "src/users/entities/user.entity";
@@ -11,12 +12,13 @@ import { Report } from "./report.entity";
 import { ReportSubmissionData } from "./report.submission.data.entity";
 
 @Entity()
+@Index(["submittedAt"]) 
 export class ReportSubmission {
   @PrimaryGeneratedColumn()
   id: number;
 
   @Column({ type: "timestamp", default: () => "CURRENT_TIMESTAMP" })
-  submittedAt: Date; //@TODO Add a key
+  submittedAt: Date;  
 
   @ManyToOne(() => User, (user) => user.reportSubmissions)
   @JoinColumn()
@@ -25,7 +27,7 @@ export class ReportSubmission {
   @ManyToOne(() => Report, (report) => report.submissions)
   report: Report;
 
-  @Column({ type: "jsonb" })
+  @Column({ type: "jsonb", nullable: true  })
   data: Record<string, any>;
 
   @OneToMany(
