+++ src/users/dto/user.dto.ts
@@ -12,3 +12,9 @@ export class UserListDto extends UserPermissions {
   roles: string[];
   isActive: boolean;
 }
+
+export class UserDto {
+  id: number;
+  username: string;
+  isActive: boolean;
+}
